% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfpca.R
\name{sfpca}
\alias{sfpca}
\title{Compute rank 1 sparse and functional principal components}
\usage{
sfpca(X, lambda_u = 0, lambda_v = 0, alpha_u = 0, alpha_v = 0,
  Omega_u = diag(nrow(X)), Omega_v = diag(ncol(X)))
}
\arguments{
\item{X}{A data matrix. Considered n x p by convention.}

\item{lambda_u}{Sparsity parameter for left singular vectors.}

\item{lambda_v}{Sparsity parameter for right singular vectors.}

\item{alpha_u}{Smoothness parameter for left singular vectors.}

\item{alpha_v}{Smoothness parameter for right singular vectors.}

\item{Omega_u}{Roughness penalty matrix (n x n) for left singular vectors.
Must be positive semi-definite.}

\item{Omega_v}{Roughness penalty matrix (p x p) for right singular vectors.
Must be positive semi-definite.}
}
\value{
List with three named elements: left singular vector u, eigenval d,
  and right singular vector v.
}
\description{
Compute rank 1 sparse and functional principal components
}
